MBS-4618: Get image uploads to go into the correct position
The image ordering code seems to be extremely brittle, so I have made some
database modifications to force ordering to be gapless. Furthermore, there
seemed to be a bug in the add cover art form which assumed that the starting
position was always 1, regardless of the amount of existing images. This has now
been changed to use count(images) + 1 as the initial position, that is - the end
of the list of images.