MBS-4640: Remove ThickBox logic for displaying web pages, we only need it for images
MBS-4640 is a bug where viewing the /front image fails. This happens because
ThickBox thinks that as this URL doesn't have .jpg, that we are probably
intending to display a HTML fragment. We don't need this functionality (and it's
wrong), so I've just remove all of that logic entirely. It now always trys to
display images.