MBS-4662: Move /relationship/attributes to /relationship-attributes
/relationship/attributes/{gid}/edit does not work, because Catalyst assumes that
'attributes' is a type of relationship type, like 'artist-artist'. I've fixed
this overlap by moving relationship attributes to live in the
/relationship-attributes namespace.

On reading this code I've taken the liberty to simplify it to use the 'Load'
role for loading, and also tested that editing works (preventing regressions)
and also that a few other links still work as I broke them while developing
this.