MBS-4687: Rewrite the Relationship::LinkType controller to use Role::Load and not have overlapping URLs
Like the move for LinkAttributeTypes, LinkType has now been simplified to use
Role::Load. I've added a separate chained action for sypecific types of
relationships - such as artist-artist or label-work. This chained action handles
validation, to disallow people trying to create moose-pony relationships, for
example.

The form has been changed to not directly depend on values in $c->stash, that's
very ugly. Instead, the link type tree is now a required attribute on the form
and must be explicitly passed in.

Tests have been added to make sure all links work correctly, along with a test
for the broken 'create' link.