Create all edits in the website within a transaction
MBS-4088 identified a terrible bug where some of the edits in the release editor
are entered, but some are not (such as ones that crash). The correct behavior is
for all of these to fail, but this wasn't happening because the transaction was
too short.

I've moved the transaction outside creating edits, which has meant moving it up
into controllers.

As this is quite a wide reaching change, I've had to create tests for all
controllers to guarantee everything is working OK,