Fix CAA indexing for indexing trigger by release label changes
Because the metadata.xml needs to be updated when a release's labels/catalog
numbers change, there is a trigger to kick off reindexing. However, the trigger
had an incorrect predicate where it tried to lookup releases by the
release_label id.

I've also added a join on cover art, so we only index stuff that actually has
cover art.