Fix unique primary aliases per locale
The original trigger was intended to set the primary column to 'false' when a
new primary alias was added for that artist, with a given locale, but due to a
bug it was setting primary to false for *all* aliases with that locale, over all
artists.

The trigger has been fixed now, and verified correct with pg_prove (see
t/pgtap). Sadly the trigger functions have got a little more complicated due to
the somewhat restrictive nature of plpgsql.