Factor all clean up scripts into a common 'RemoveEmpty' script.
This script can be invoked with an entity type, and it will remove all entities
of that type. This reduces the duplication between the cleanup scripts. I've
moved all SQL logic to stored procedures, which has meant writing empty_works()
as this was previously done in Perl code.

Also added tests for empty_artists()