MBS-3904: Do not allow yes/no/abstain votes on own edits
The 'approve' vote needs to be allowed, as this is what happens when an
auto-editor approves their own open edit. However, manually crafting a form
submission will only enter the vote but won't actually do the approval, nor does
the edit queue use 'approve' votes when making decisions, so being able to enter
this vote is not harmful.