MBS-4382, Display release length.
This commit will display the release length in the sidebar IF the data
has been loaded from the database.  In practice that means the
"Release" tab shows the duration if all tracks have a length, and the
"Disc IDs" tab shows the duration if all discs have a discid.
Entity::Release->length() will use whatever is available, so we can
simply load the tracklists and/or cdtocs if we want to show the
duration elsewhere.