MBS-2662: Fix the DISTINCT clause of a few reports
Simply using DISTINCT is not enough, because we include an integer that
indicates the sort order. Thus if you have multiple rows, this integer will
always be unique and thus DISTINCT is fairly pointless. I've moved some queries
to subselects that do have DISTINCT, and then done the ordering outside that
subselect, which only use n:1 joins.