MBS-3590: Use memcached to atomically generate wizard session IDs
It appears that the session ID is not stable, and is prone to change. From
debugging this issue on IRC, we found that of all 3 wizard_session_id's that
resulted in this crash, there were memcached keys with different session IDs,
but the same wizard ID. It seems unlikely that this is due to a clash in rand()
3 times in less than an hour.

I've changed the wizard session ID to be generated by memcached, providing a
global, and stable incremental ID store.