Fix editor test to find now() by the database not DateTime
The update email test updates an editor's confirmed email by setting the date to
now() - the time the parent transaction started. As we run all tests inside a
transaction, DateTime->now could actually be after the database's idea of now(),
which leads to the test failing as it thinks the confirmation date is set to
something in the past.

To work around that, I've just set $now to be the database's idea of now().