Adjust 'edits_pending' columns before creating edit_* links
The insertion of an edit_* link row will acquire a share row lock on the
referenced row (ie, the artist row). Later updating this row to change the value
of edits_pending will require a more heavy weight lock. However, if a parallel
transaction has also created a edit_* row, then that will block the exclusive
lock. This parallel session will the deadlock against the first as it too wants
an exclusive lock.

By doing the updates first, we end up acquiring a more heavy weight lock
throughout the transaction.