MBH-142: Cache /text.js
/text.js is now accessed with a query parameter of the current checked out Git
revision. It also also served with a large expiration date (unless the server is
in a development environment). For people in an environment where URLs with
query strings are cached, this will yield a roughly 200ms reduction in page load
time. For other users, there will be no change (for better or worse).