MBS-4930: Establish a fresh database connection on every request
If a sql statement throws an exception (for example, in the case of a deadlock)
then the transaction is aborted, but will not be rolled back. Furthermore, if
an exception is thrown it will be rethrown in Perl code, which sometimes means
that the whole request fails and the transaction is never rolled back. Thus, on
subsequent requests we begin with an aborted transaction, and the server is in
an inconsistent state.

There are a few 'correct' solutions. One is to ensure that all SQL calls that
can fail are bracketted in a transaction, but this leaves room for bugs such
as this one to remain in the code. Another would be to just roll back at the
start or end of requests, but this either risks not being called, or producing
unnecessary noise in PostgreSQL logs (rollback when not in transaction).

By connecting on startup we incur a little overhead, but I plan to use
pgbouncer to mitigate this, on totoro.