MBS-3957: Check for a differences track durations
Check differences in track duration changes, rather than a difference between
track and recording duration. This means that even if a track has a duration
but a recording does not, as long as the track duration has not changed from
what was already in the database the editor will not have to confirm the
recording.

If an editor does change a track duration by >10s then they will have to
confirm the recording association.