MBS-5091: Apply edits from FixTrackLength inside transactions
The edits that are created in FixTrackLength are intended to be created from
within transactions. At the moment, they are not, which means that one edit
will successfully be inserted, but the subsequent update fails (with a fatal
exception) and the script terminates. This has lead to a lot of
"open auto-edits" created by ModBot.