MBS-4354: Correct the display of add isrcs edit when client is not present
If a web service client submits an add ISRCs edit but does not provide client
information, the client will default to the empty string. This is defined, to
differentiate it from the actual web site, which uses undefined. If an empty
string is encountered when display an edit, we instead display (unknown).