Submit the relationship editor form using ajax
Data is now posted to /relationship-editor using ajax. If there are errors,
it returns them as JSON. This means the page no longer has to refresh, and
the previous code to reconstruct relationships from the CGI params has been
removed.

Changes have also been made to Controller::RelationshipEditor to make it more
re-usable. e.g. if we wanted to add /recording/mbid/edit-relationships, it's
as simple as $c->forward('/relationship_editor/load') in the recording
controller (ideally).