MBS-5124: Only create LWP in a single place and correctly set timeout
We were creating various different LWP objects with different settings, and in
some cases setting properties incorrectly. The worst case of this was in the
/ws/2 xml_search routine, which would set the timeout *after* making the
request!

I've added an LWP attribute to the context, so everything can make use of a
single LWP object, and we have a central place for configuration.