MBS-5110: Serialize tagging entities based on entity ID
If 2 threads are running the transaction at the same time, with the same input
they will usually fail because the first transaction will sucessfully insert the
tag, while the second transaction will fail due to violating the primary key. By
serializing transactions on the entity themselves, this violation is no longer
possible.