MBS-2650: Redirect to the relationships/overview page when adding relationships
When a user creates a relationship from a source entity A, once they have
finished creating the relationship we should redirect them back to a specific
page of A which shows the most amount of information about relationships. For
most entities this is the /relationships page, but for recordings, URLs and
works, this is simply their landing page.