MBS-5177: Do not allow editors without email addresses to edit
If an editor removes their email address then they should immediately be denied
any further editing. This patch changes the data stored in the session from a
partial editor object, to simply the editor ID. By caching the user in the
session we are forcing ourselves to maintain 2 stores of data, and we're doing
a bad job of this. Rather than using, a cache, just fetch editors from the
database per request.

This will also mean that if an admin removes a users email address, then the
change will propogate immediately to the editor in question.