MBS-3360: Add scripts to add the new functions and remove unused tags
There are a few indexes that were missing, which will be needed to speed up
deletions on the tag table. The cleanup script itself simply does a sequential
scan of the entire 'tag' table, trying to delete each row using the
delete_unused_tag function (thus only deleting rows that can be deleted).