MBS-1579: Add a 'swap track titles with artist credits' button to release editor
On click, this button will swap the 'artist' column contents with the 'title'
column, and vice versa. If the artist credit for any of the tracks in the
tracklist is deemed 'complex', then the user is asked to confirm the action. A
complex artist credit is one that has a confirmed artist ID, or multiple artist
credit parts. This is because swapping track titles and artist credits discards
this information, and it can't be undone.