MBS-2650: Conditionally include sidebars
Everytime we are going to introduce a sidebar, we first check whether full_width
is true or false. If it is true, then we do not process the
sidebar. layout/sidebar.tt does have this logic, but this logic is only applied
when deciding whether or not to show the sidebar content - the sidebar template
is still evaluated first. By moving the condition up to before the WRAPPER call,
we don't have to evaluate the template at all.

This was introduced to resolve MBS-5277, which is caused by the changes to
common-macros.tt in c233a64c48fd6db.