MBS-5250: Improve 'has_age' determination
has_age has been changed to be the conjunction of:

* The begin date has at least a defined year field.
* The begin date is non-negative.
* The begin date is before now().
* The end date is empty OR the end date is after the begin date.

The latter is determined by walking both the begin and end date's defined
components (that is, all date elements that are *not* undefined, stopping at the
first value that is undefined) and comparing them. For the end date to be after
the begin date, an end date field must be greater than a begin date field,
before a begin date field is greater than an end date field.