MBS-5250: Add take_while utility function
Adds the 'take_while' function which when given a list A and a predicate,
returns a new list with all elements of A where the predicate held true, until
the predicate held false.

I encountered a problem exporting this function, and found out that we don't
use strictures for Data::Utils at all. Turning this on introduced a few new
problems, so I've had to tidy them up. Specifically, I've fixed some invalid
syntax and removed the circular dependency.