MBS-3680: Replace /ws/2 XML serializer with an event based XML::Genx impl.
The previous implementation first built up a DOM of MusicBrainz::XML type nodes,
and then finally collapses this into a string by walking the DOM. Thus, we end
up recursing over 2 data structures and allocating, when we could do everything
in one go. That's what this patch does.

The new serializer uses XML::Genx to emit events when it needs to render. So
instead of building up another structure, we just render XML as we go. This
roughly halves rendering time for large documents. The downside is that this is
more reliant on side-effects, the order of function calls and it means we won't
be able to memoize any of these calls. I don't expect us to stay with Perl for
much longer (or continue to have such large documents) that these are acceptable
trade-offs.