MBS-5344: Clarify 'we cannot display artwork' message for deleted releases
The message informing the user that we cannot display history for artwork has
been changed to not link to a release page, if the release itself no longer
exists. I've also made a minor change to indicate that we can't display history
for *this* cover art, because we're obviously showing some history for some
cover art!