MBS-3853; Fix annotation history JavaScript
There was a weird little bug which meant that every other row caused the wrong
radio buttons to appear. This was due to the fact that `false || undefined ==
undefined`, and .attr('checkbox') would return undefined if that attribute
wasn't present. I've used the !! operator to coalesce undefined to false, and
now the control displays as intended.