MBS-2966: Revert adding asterisk for name variations.
The reporter of the ticket proposed to simply add a class to entity
links with name variations instead of specifically marking them with
e.g. "*". This revert keeps the class name "name-variation" for these
links, allowing user script/style authors to manipulate their display
and/or behaviour to their liking.