Allow page timeouts to be configured per page, rather than per app
MAX_REQUEST_TIME is now DETERMINE_MAX_REQUEST_TIME, which is a function from
a Catalyst::Request to a duration in a seconds, that defines the amount of
time it can take for a response to be generated. This allows us to match
the request path to timeout specific pages at a different rate than others.