MBS-5698: Display historic 'move disc ID' edits correctly
When the migration script ran, it attempt to resolve a release ID into the set
of release IDs that the release was migrated to in NES (as multiple release
events produced a new release ID). However, if this resolution couldn't be done
because the release was deleted since the edit, the result was the empty set.

When displaying the from/to release set in the edit display, we now check for
whether or not the release set is the empty set and if so display
'[removed]'. Otherwsie, we show all the releases.