MBS-5788, MBS-5791: Fix creation of move disc ID/remove ISRC edits (and more)
These edits called $c->detach inside their on_creation hooks, which causes an
exception to be thrown. MBS-5586 (26eee) moved these hooks to be inside the
transaction, and so when the detach exception is thrown the transaction is
first rolled back.

This commit stops calling $c->detach inside these handlers, and instead performs
this outside the transaction.