MBS-5797: Don't crash on non-Wikipedia Wikipedia relationships
Controller::Role::WikipediaExtract is used to add Wikipedia summaries on an
entity, based on Wikipedia URL relationships. However, if a Wikipedia URL
relationship is present, but the URL is not actually to Wikipedia, an exception
would be thrown.

This commit addresses that by filtering the list of relationships considered to
only be those relationships that truly do point to Wikipedia.