MBS-5838: Fix format_edit_note to preserve whitespace around links
Add more tests to MusicBrainz::Server::Filters::format_edit_note to make sure
that if a link follows some plain text and whitespace, the text and whitespace
is preserved. 95c57dbcf added a schema-independent URLs, but in doing so would
strip any leading whitespace before the URL. I've now changed the starting URL
boundary to be either the start of the edit note, or a negative look behind for
':'.