MBS-3856: Rewrite the query and remove the run time limit
The query now does an array aggregation of all URL link types against a release,
and checks if that array overlaps with the array of link types that can provide
cover art. If there is no overlap, then the release_coverart row is regenerated.

The run time limit has also been removed, as it serves no real purpose (the
transaction boundary is per release).