MBS-5873: Correctly link label conflicts to label pages
My fix for MBS-5696 made ModBot leave comments when an edit would
cause a violation of duplicate rows, but I incorrectly hard-coded
the URL path for artists, even if the error was thrown for labels.
This commit moves the URL generation to subclasses, and changes
the wording to mention the more general 'entity', rather than
artist or label.