MBS-5873: Do not suggest merging entities into themselves
Edit 20758849 suggested that the editor merge a label into itself,
which is clearly not possible. It did this because a subsequent edit
was applied, which did applied the same change as that proposed by
the edit in question. This meant that when the edit was applied,
it would violate unique keys, because it didn't exclude itself.

This commit adds a check into the uniqueness assertions to *not*
consider the entity we are checking.