Rename setup_test_db_for_selenium.pl → script/setup_development_db.pl.
Now that selenium tests have been removed, a script to set up a
database for selenium is no longer needed.  However, the script is
still useful as a general way to set up a minimal database.  I changed
it to operate on the production/development database instead of a test
database.