MBS-357: Add a long running process to bcrypt passwords
bcrypt-passwords.pl listens on the `EditorChanges` PgQ for events that indicate
that an editors plain text password should be hashing using bcrypt. Triggers on
`INSERT` and `UPDATE` to the `editor` table cause passwords to rehash.

If the password is `NULL` or the empty string, it's assumed that this means
logins shouldn't be allowed on this account, in which case the `RejectAll`
authenticator is used.