MBS-6021: Use `tracksmedium` field to find discs in release editor
The `tracksmedium` field finds releases by the amount of tracks that appear on
one medium, whereas the previous `tracks` field would search the track count of
the entire release (the sum of all tracks). Users expect to search by the tracks
on one medium, as they are adding single discs.