MBS-6058: New schema to model deleted/merged entities separately
This new schema allows us to have real foreign key constraints for active
subscriptions, as the bug in MBS-6058 occurs because HumHumXX is subscribed to
artists that don't exist (but also have apparently not been deleted).

The _deleted tables point to artist_/label_deleted rows, which collect lists of
MBIDs that have been deleted, along with the last known name/comment in order
for us to display the deletion clearly in the subscriptions. This data is
probably interesting to customers too, and has uses for us internally (better
404 pages).