Implement a mandatory password reset for all editors
See http://blog.musicbrainz.org/?p=1844 for why we need this.

Checks the `last_login_date` for editors on login, and if that's before
`$LATEST_SECURITY_VULNERABILITY` they are immediately logged out and forwarded
to the "change password". This page has been updated to work when people are
logged out, requiring they enter a username. The username is prefilled from the
currently logged in user's name, or can be filled in from query parameters. The
login form uses this to fill it in for a better user experience.