MBS-6023, Remove dead code in release editor.
As far as I can tell these particular code paths are never executed on
production.  The $look_for_tracklist is always set to 0 immediatly,
because an artist credit doesn't have a ->{id} (the artist has the id,
so ->{artist}->{id}).

The call to model ('ArtistCredit')->find (@artist_joinphrase) is
further proof that this code never runs -- find() would die with an
exception as it expects a single artist credit hashref (with a
->{names} key), not an array of names.

This seems to have never worked, most of this code is unchanged since
march 2011, so this commit just removes all of it -- this as far as I
can tell will not change the behaviour compared to current production.