MBS-6023, fix release editor, add disc -> existing tracklist for track identifier changes.
Before we got rid of tracklists, the recording associations tab could
suggest the recordings associated with a tracklist when the "existing
tracklist" feature was used to add a disc to a release.

But "existing tracklist" cannot set the medium_id for the disc to an
existing medium, because that would mean the medium would get attached
to multiple releases.  To communicate from which medium the tracks
were loaded a "medium_id_for_recordings" field was added to the
tracklist form, it is read by the recordings tab to associate
recordings with those tracks.