Fix /ws/2/json tests
This 'fix' forces the ordering of relation elements in the JSON response which
means that the actual response is consistent with the expected result.

I've also changed some 'TRUNCATE' commands to 'DELETE FROM', as these are faster
to execute.

Finally, there was a bizzare caching problem which was caused by
MusicBrainz::Server loading before 'prepare_test_server' was called, which
resulted in an in memory cache being used for the tests.