MBS-6326: Don't use country_id in edits
This pull request stops using 'country_id' in
Edit::(Artist|Label)::(Create|Edit) edits. If this data is present then the edit
is assumed to come from an old edit that is still open, and we use an overriden
'restore' implementation to move this field into the 'area_id' field.