MBS-6317: Allow loading EditReleaseLabel edits
Some EditReleaseLabel edits are unable to load because they are trying to create
a PartialDate using the empty string as a single parameter to the constructor.
However, this fails to pass the 'is a well formatted' string predicate in
PartialDate, and thus the PartialDate constructor tries to use the string as a
hash reference.

The empty string is a well formatted partial date string, so the predicate has
been relaxed to accomodate for this case.