MBS-6340: redirect to direct search for areas
Until we have indexed search for areas, this makes it so that the main
/search page turns indexed searches into direct ones (like for editor
and tag), and autocomplete boxes will default to direct search for
areas. It's still possible to manually choose to switch to indexed
search in autocompletes, because this is intended as a quick patch for
the hopefully short time until we have area search for real.