MBS-6275: Attempt to correlate changed release events in edits
Now tries to match old and new release events that have the same
country. In the case that they do, we use a full character diff to
display the exact differences. For remaining events that don't have
a common country on both sides of the diff, we list all countries.