MBS-6471: Consume 'remember me' tokens in a 5 minute window
This commit changes the 'remember me' system to use Redis as the storage, rather
than PostgreSQL. This allows us to store remember me tokens with expiry dates,
and thus enables us to have a grace period for remember me token consumption.
This is needed to solve the problems in MBS-6471.

MBS-6471 describes a problem where users are getting logged out, even though
they have ticked 'remember me'. I believe this happens if multiple requests are
sent to musicbrainz.org *without* session cookies. This means one request will
consume the token, while subsequent requests will fail and won't log in. The
latest request sets the session, so that will usually be one that couldn't log
in.

Instead, we now consume tokens by setting an expiry time of 5 minutes in the
future, rather than immediately consuming the key. This means that if a burst of
tabs are opened, all of them can use the same remember me token.

Tokens are also initially created with a 1 year expiration period, in order to
put a weak guaranntee on the amount of space Redis will require.