MBS-6486: ModBot should remove collaboration relationships on AC split
f0b7974de504d3b07 changed the call to edit_action for Artist->split to have to
entries for on_creation. As such, the latest definition wins (because the
arguments to on_creation are a mapping and there can only be one entry per key).

This commit merges both on_creation entries into a single entry, and also adds a
test to ensure that behaviour is correct.