Move MB::S::get_collator to Translation, fix Form::Area test
The Form::Area test builds a custom testing context so it doesn't have to bring
up the entire Catalyst stack. However, 83afe0f5 added a new method to the
Catalyst context, which won't be shared with this new form context. I've
factored out get_collator to MusicBrainz::Server::Translation so I can at least
provide this in the test context.