MBS-6384: Serialize CD stubs in JSON
The /ws/2/discid end point will return CD stubs if MusicBrainz does not have a
real cdtoc for the disc ID passed. For XML this is fine, as there is a
serialiser defined for CD stubs, but JSON does not have a serialiser.

First I improved the error message about missing serialisers to be a little bit
more informative, and added tests for that. Next I added a CDStubTOC serialiser
that does the same as the XML serialiser, along with tests for just this
serialiser.