MBS-6426: Do not destroy release events when editing other release information
The merge implementation would always set $merged->{events} to $new->{events},
even if this key didn't exist in $new. This would leave $merged->{events} as
`undef`, but unfortunately autovivification means that this gets inflated to []
in the `grep` introduced in the parent commit.  As such, Release->update would
interpret this as clearing all release events for a release.

This commit displays setting $merged->{events} in this case.