MBH-334: Attempt to reconcile differences against production DB
The production database has a lot of temporary tables that shouldn't
be there, so I've dropped those. Various check constraints don't have
names that match up with what's in the codebase and while that's logically
sound, it's still a pain to do comparisons.

This commit takes a first attempt at resolving MBH-334 by picking
the low hanging fruit.<