MBS-6548: Partial merge of MBS-5487 for ModBot
MBS-5487 has been merged to beta, which means that edits created on beta have a
schema that differs from the one that ModBot (via astro) will see. This means
that we have a large amount of edits failing to apply every hour. This makes it
harder to spot real problems, and makes the voting queue difficult to navigate,
so in this commit I am merging the library changes of MBS-5487, without exposing
the UI.