Require editor's birth year after 19th century
Ensures that the 'birth date' field for editors contains values on
or after 1900-01-01. This does make an assumption that the person
behind any account was born in (or after) the 20th century, but I
think this is a reasonable assumption to make. There is a tiny chance
that someone lived to be over hundred years old at the start of the
project, but I think this is so rare I have chosen to go with 1900
as this makes the error message look less arbitrary (than say, 1850).