MBS-6360: Respond to non-UTF8 authenticated web service requests with 400
Catalyst::Plugin::Unicode::Encoding usually deals with all of this for us,
however this plugin only tries to deal with query/body parameters and URL
captures. For digest authorization, none of these are used, and instead we look
at the Authorization header. Catalyst::Plugin::Authentication::Credential::HTTP
doesn't try and do any decoding on this parameter, thus it passes the string
directly through to our Data:: layer, and finally PostgreSQL complains as it
expects everything to be valid UTF-8.

In this commit I've added a small check before we drop into Catalyst
authentication, which simply ensures that the Authorization header does form a
valid UTF-8 encoding of something. If not, the request is immediately responded
to with a 400 response.