MBS-6570: Allow clearing relationship dates in relationship editor
If a begin or end date are removed for a relationship, the field doesn't
get submitted at all. To the server code, this is interpreted as
"the user doesn't want to modify this field, whatever it's current contents
are" - which is not what the user intendend.

As a workaround, I am now defaulting undef begin/end dates to an empty
hash, in the code that submits "edit relationship" edits. An empty hash
means unset year/month/day fields.