MBS-3132: Correctly set track artist credits from release artist credits
There was a small bug in the 'add' version of the release editor,
whereby if the user seeded the whole release editor with a single artist
and later changed the release artist to an artist credit consisting of
multiple artist credits, only the first artist credit would be used
for tracks.

This happens because some of the individial artist credits in the
release artist credit do not have the 'name' attribute set. This causes
ArtistCredit->from_array to skip them entirely, thus only those with
that attribute are used (which happens to only be the first artist credit).

Using clean_submitted_artist_credit causes this attribute to be
populated from the nested '{artist}->{name}' attribute, and now all
artist credits are used when defaulting the track artist credits to that
of the release.