MBS-6578: Make area merges possible when areas are used by entities
If an area was used by an artist or label as the begin, end, or general
area (outside relationships), then the area itself cannot be merged.
This is because the deletion of the old area may now violate foreign
key constraints.

In this commit I move over these references to the new area, if an
entity has a not-null reference to one of the old areas.