MBS-6578: Make area merges possible when areas are country areas
If an area is a country area, then there is a corresponding row
for that area in country_area. However, Area->merge does not
consider these rows when merging, so it is possible to violate
foreign key constraints on merge.

This commit considers these rows, by making the destination area
a country_area, if any of the old areas are country_areas. It then
moves all release_country references to point to the new country_area
and deletes the old ones.