MBS-6579: Do not use RDFa when showing release event diffs
Using RDFa means that our diff code tries to diff changes in
HTML attributes. This leads to invalid HTML, in a few cases
(such as adding a month to a release date that was only a year).
To work around this, I've just turned RDFa off entirely in
the 'edit release' edit pages.