MBS-6496: Do not use medium recordings when adding new discs in release editor
When a new disc is added in the release editor, the last disc in the
editor is cloned and all its fields are cleared. However, there
appears to be a slight mistake in this clearing - the 'id' field is
cleared twice.

I've changed the second reference to the 'id' field to
'medium_id_for_recordings', which is an id that is used to provide
recording association suggestions. Clearing this, the release
editor suggests that new recordings are created.