MBS-6605: Fix the display of /isrc pages
By default, Controller::Role::Load will try to set
`$c->stash->{model_to_type($model)}` to the result of `_load`,
if `model_to_type($model)` is a valid mapping.

Previously, this wouldn't resolve to anything, and so this key
would remain unchanged. However, d234581098ebdd added mappings for
ISRCs and ISWCs, and so after `_load`, `$c->stash->{isrc}` was set to
the result of `_load` (which amusingly, was `$c->stash` itself).

In this commit, I have added another condition that must be met
for this default setting to occur - the absence of any value in
this key.