MBS-4669: Add a fallback to populate release_meta
There are some Amazon ASIN's which are valid, but do not contain
any artwork. If a release has only a relationship to an Amazon ASIN
of this nature, then release_meta should still contain the ASIN, but
not link to any artwork.

The cover art code previously only populated release_meta from the
results of a sucessful cover art lookup - if cover art lookup failed,
then there would be no ASIN to populate release_meta with.

This commit introduces the idea of 'fallback_meta', which is used
to populate release_meta in the scenario that no artwork is found.