MBS-6628: Correctly accept merge label edits with subscriptions
The Edit::Role::MergeSubscription role was incorrectly generalised, and
still had 'artist' hard-coded. This works fine for artist merges, but
fails for label merges as the deleted subscriptions have the 'label' key,
rather than 'artist'. This commit changes this code to call model_to_type
to identify the key name.