MBS-4788: Disable all release group type fields when using existing release group
If the user has selected a release group in the release editor, they should be
unable to also edit the type of the release group (primary or secondary). The
code that was previously responsible for this was not updated for multiple
release group types, so I have updated it to now (correctly) consider the
primary release group type, but also all secondary release group types.