MBS-6644: Add --limit to ProcessReplicationChanges
This flag indicates how many rows ProcessReplicationChanges should
fetch from dbmirror_pendingdata when applying the changes in a transaction.
The limiting is implemented by fetching rows via a cursor, using
the existing query.

The default value is 10000 which causes ProcessReplicationChanges
to use 32MB max. resident memory when applying 90k updates to
area_type (which admittedly is fairly arbitrary tuning, but the
majority of packets do not have transactions anywhere near this size).