MBS-6529: Fix display of edit release label edits entered before schema 18
The old edit schema was to have 'release.country' which contained
a string - the name of the country. With schema 18, this property
went away in favour of release.events.country_id. The remapping code
attempted to rewrite 'release.country' to 'release.events.0.country_id' -
but 'release.country' is not an integer so this doesn't make sense.

This commit changes the mapping to instead consult
Area->search_by_names to determine the country_id, and if this
fails it uses the new country_name field (if that is set).