MBS-6656: relationship editor can't count
The number of selected recordings wasn't being counted properly ever since
the MB.entity changes.

When knockout is rendering a recording on the page, it's actually using a
Track entity for the binding context. To get the checked recordings entities,
we were selecting the checked inputs with jQuery, and then running ko.dataFor
on each input (which returns the data bound to that element). To fix this bug,
I wrapped every recording checkbox inside a with-binding, so that the binding
context for the checkbox is a recording instead of a track.

Also in this patch are a few tiny simplifications to the code.